[Geomagic plugin] reorder plugin files for better modularisation
