+++ applications/plugins/Geomagic/src/Geomagic/GeomagicDriver.cpp
@@ -20,7 +20,7 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 
-#include "GeomagicDriver.h"
+#include <Geomagic/GeomagicDriver.h>
 #include <sofa/core/ObjectFactory.h>
 #include <sofa/simulation/AnimateBeginEvent.h>
 #include <sofa/simulation/AnimateEndEvent.h>
